[feature/ms-303][Picking delivery] + release picking and delivery ver 0.3

+++ apps/masterdata/saledata/models/good_receipt.py
@@ -109,7 +109,7 @@ class GoodReceiptProduct(MasterDataAbstractModel):
             }
         ),
     )
-    quantity = models.IntegerField(
+    quantity = models.FloatField(
         verbose_name="Quantity",
         help_text="quantity of product",
     )
